First version of the Philips Hue Lua API.
I'm a Lua newbie so I still haven't figured out how to best package
this as a library for the community, but for now, at least the code is
online. This includes a simple Hue API (hue.lua) and a simple command
line tool (huectl) that can be used to control your Philips Hue
lights.