Normalized argument names and added ability to get state of lights
Argument names that set state are now named as they are in the Hue
SDK. This allows me to point folks to the SDK documentation for
information on those settings.

Added the ability to retrieve the state of lights or the last action
sent to a group. One can retrieve all of the state or just a component
of the state using the -g option.